Removed unnecessary Exception classes and the uses of these classes from the resource classes.

+++ midolmanj-mgmt/src/main/java/com/midokura/midolman/mgmt/rest_api/resources/VifResource.java
@@ -1,7 +1,6 @@
 /*
- * @(#)VifResource        1.6 11/09/24
- *
  * Copyright 2011 Midokura KK
+ * Copyright 2012 Midokura PTE LTD.
  */
 package com.midokura.midolman.mgmt.rest_api.resources;
 
@@ -26,22 +25,19 @@ import org.slf4j.LoggerFactory;
 
 import com.midokura.midolman.mgmt.auth.AuthAction;
 import com.midokura.midolman.mgmt.auth.Authorizer;
-import com.midokura.midolman.mgmt.auth.UnauthorizedException;
 import com.midokura.midolman.mgmt.data.DaoFactory;
 import com.midokura.midolman.mgmt.data.dao.VifDao;
 import com.midokura.midolman.mgmt.data.dto.UriResource;
 import com.midokura.midolman.mgmt.data.dto.Vif;
 import com.midokura.midolman.mgmt.rest_api.core.ResourceUriBuilder;
 import com.midokura.midolman.mgmt.rest_api.core.VendorMediaType;
-import com.midokura.midolman.mgmt.rest_api.jaxrs.UnknownRestApiException;
+import com.midokura.midolman.mgmt.rest_api.jaxrs.BadRequestHttpException;
+import com.midokura.midolman.mgmt.rest_api.jaxrs.ForbiddenHttpException;
 import com.midokura.midolman.state.NoStatePathException;
 import com.midokura.midolman.state.StateAccessException;
 
 /**
  * Root resource class for VIFs.
- *
- * @version 1.6 24 Sept 2011
- * @author Ryu Ishimoto
  */
 public class VifResource {
     /*
@@ -64,8 +60,6 @@ public class VifResource {
      *            Authorizer object.
      * @throws StateAccessException
      *             Data access error.
-     * @throws UnauthorizedException
-     *             Authentication/authorization error.
      * @returns Response object with 201 status code set if successful.
      */
     @POST
@@ -73,31 +67,20 @@ public class VifResource {
             MediaType.APPLICATION_JSON })
     public Response create(Vif vif, @Context UriInfo uriInfo,
             @Context SecurityContext context, @Context DaoFactory daoFactory,
-            @Context Authorizer authorizer) throws StateAccessException,
-            UnauthorizedException {
+            @Context Authorizer authorizer) throws StateAccessException {
+
         if (vif.getPortId() == null) {
-            throw new IllegalArgumentException("Port ID is missing");
+            throw new BadRequestHttpException("Port ID is missing.");
+        }
+
+        if (!authorizer.vifAuthorized(context, AuthAction.WRITE,
+                vif.getPortId())) {
+            throw new ForbiddenHttpException(
+                    "Not authorized to plug VIF to this port.");
         }
 
         VifDao dao = daoFactory.getVifDao();
-        UUID id = null;
-        try {
-            if (!authorizer.vifAuthorized(context, AuthAction.WRITE,
-                    vif.getPortId())) {
-                throw new UnauthorizedException(
-                        "Not authorized to plug VIF to this port.");
-            }
-            id = dao.create(vif);
-        } catch (StateAccessException e) {
-            log.error("StateAccessException error.");
-            throw e;
-        } catch (UnauthorizedException e) {
-            log.error("UnauthorizedException error.");
-            throw e;
-        } catch (Exception e) {
-            log.error("Unhandled error.");
-            throw new UnknownRestApiException(e);
-        }
+        UUID id = dao.create(vif);
         return Response.created(
                 ResourceUriBuilder.getVif(uriInfo.getBaseUri(), id)).build();
     }
@@ -115,37 +98,28 @@ public class VifResource {
      *            Authorizer object.
      * @throws StateAccessException
      *             Data access error.
-     * @throws UnauthorizedException
-     *             Authentication/authorization error.
      */
     @DELETE
     @Path("{id}")
     public void delete(@PathParam("id") UUID id,
             @Context SecurityContext context, @Context DaoFactory daoFactory,
-            @Context Authorizer authorizer) throws StateAccessException,
-            UnauthorizedException {
+            @Context Authorizer authorizer) throws StateAccessException {
 
-        VifDao dao = daoFactory.getVifDao();
         try {
+            VifDao dao = daoFactory.getVifDao();
             Vif vif = dao.get(id);
-            if (!authorizer.vifAuthorized(context, AuthAction.WRITE,
-                    vif.getPortId())) {
-                throw new UnauthorizedException(
-                        "Not authorized to unplug VIF to this port.");
+            if (vif != null) {
+                if (!authorizer.vifAuthorized(context, AuthAction.WRITE,
+                        vif.getPortId())) {
+                    throw new ForbiddenHttpException(
+                            "Not authorized to unplug VIF to this port.");
+                }
+
+                dao.delete(id);
             }
-            dao.delete(id);
         } catch (NoStatePathException e) {
             // Deleting a non-existing record is OK.
             log.warn("The resource does not exist", e);
-        } catch (StateAccessException e) {
-            log.error("StateAccessException error.");
-            throw e;
-        } catch (UnauthorizedException e) {
-            log.error("UnauthorizedException error.");
-            throw e;
-        } catch (Exception e) {
-            log.error("Unhandled error.");
-            throw new UnknownRestApiException(e);
         }
     }
 
@@ -164,8 +138,6 @@ public class VifResource {
      *            Authorizer object.
      * @throws StateAccessException
      *             Data access error.
-     * @throws UnauthorizedException
-     *             Authentication/authorization error.
      * @return A VIF object.
      */
     @GET
@@ -174,29 +146,18 @@ public class VifResource {
             MediaType.APPLICATION_JSON })
     public Vif get(@PathParam("id") UUID id, @Context SecurityContext context,
             @Context UriInfo uriInfo, @Context DaoFactory daoFactory,
-            @Context Authorizer authorizer) throws StateAccessException,
-            UnauthorizedException {
+            @Context Authorizer authorizer) throws StateAccessException {
 
         VifDao dao = daoFactory.getVifDao();
-        Vif vif = null;
-        try {
-            vif = dao.get(id);
+        Vif vif = dao.get(id);
+        if (vif != null) {
             if (!authorizer.vifAuthorized(context, AuthAction.READ,
                     vif.getPortId())) {
-                throw new UnauthorizedException(
+                throw new ForbiddenHttpException(
                         "Not authorized to view this VIF.");
             }
-        } catch (StateAccessException e) {
-            log.error("StateAccessException error.");
-            throw e;
-        } catch (UnauthorizedException e) {
-            log.error("UnauthorizedException error.");
-            throw e;
-        } catch (Exception e) {
-            log.error("Unhandled error.");
-            throw new UnknownRestApiException(e);
+            vif.setBaseUri(uriInfo.getBaseUri());
         }
-        vif.setBaseUri(uriInfo.getBaseUri());
         return vif;
     }
 
@@ -213,8 +174,6 @@ public class VifResource {
      *            Authorizer object.
      * @throws StateAccessException
      *             Data access error.
-     * @throws UnauthorizedException
-     *             Authentication/authorization error.
      * @return A list of VIF objects.
      */
     @GET
@@ -222,30 +181,18 @@ public class VifResource {
             MediaType.APPLICATION_JSON })
     public List<Vif> list(@Context SecurityContext context,
             @Context UriInfo uriInfo, @Context DaoFactory daoFactory,
-            @Context Authorizer authorizer) throws StateAccessException,
-            UnauthorizedException {
+            @Context Authorizer authorizer) throws StateAccessException {
 
-        VifDao dao = daoFactory.getVifDao();
-        List<Vif> vifs = null;
-        try {
-            if (!authorizer.isAdmin(context)) {
-                throw new UnauthorizedException(
-                        "Not authorized to view all VIFs.");
-            }
-            vifs = dao.list();
-        } catch (StateAccessException e) {
-            log.error("StateAccessException error.");
-            throw e;
-        } catch (UnauthorizedException e) {
-            log.error("UnauthorizedException error.");
-            throw e;
-        } catch (Exception e) {
-            log.error("Unhandled error.");
-            throw new UnknownRestApiException(e);
+        if (!authorizer.isAdmin(context)) {
+            throw new ForbiddenHttpException("Not authorized to view all VIFs.");
         }
 
-        for (UriResource resource : vifs) {
-            resource.setBaseUri(uriInfo.getBaseUri());
+        VifDao dao = daoFactory.getVifDao();
+        List<Vif> vifs = dao.list();
+        if (vifs != null) {
+            for (UriResource resource : vifs) {
+                resource.setBaseUri(uriInfo.getBaseUri());
+            }
         }
         return vifs;
     }
